--- app/api/db.ts
@@ -1,12 +1,0 @@
-type Db = {
-  asteroids: AsteroidOnClient[];
-  asteroidsInCart: AsteroidOnClient[];
-  ids: string[],
-  counter: number;
-};
-export const db: Db = {
-  asteroids: [],
-  asteroidsInCart: [],
-  ids: [],
-  counter: 0
-};
